Move managers and template into subdirectories
