+++ openpassword/agile_keychain/_manager/_item_manager.py
