+++ openpassword/agile_keychain/data_source.py
@@ -6,9 +6,9 @@ from time import time
 from openpassword import abstract
 from openpassword.exceptions import KeyValidationException, IncorrectPasswordException, \
     UnauthenticatedDataSourceException
-from openpassword.agile_keychain._file_system_manager import FileSystemManager
-from openpassword.agile_keychain._key_manager import KeyManager
-from openpassword.agile_keychain._item_manager import ItemManager
+from openpassword.agile_keychain._manager._file_system_manager import FileSystemManager
+from openpassword.agile_keychain._manager._key_manager import KeyManager
+from openpassword.agile_keychain._manager._item_manager import ItemManager
 from openpassword.agile_keychain._crypto import create_key, decrypt_key, encrypt_key, \
     decrypt_item, encrypt_item, generate_id
 from openpassword.agile_keychain.agile_keychain_item import AgileKeychainItem
