+++ openpassword/agile_keychain/template/1password.keys.template
