+++ tests/integration/openpassword/agile_keychain/test_file_system_manager.py
@@ -2,7 +2,7 @@ import os
 import shutil
 from nose.tools import raises
 
-from openpassword.agile_keychain._file_system_manager import FileSystemManager
+from openpassword.agile_keychain._manager._file_system_manager import FileSystemManager
 
 AGILE_KEYCHAIN_BASE_FILES = ['1password.keys', 'contents.js', 'encryptionKeys.js']
